Fix more compile warnings

+++ src/DocHistory.cpp
@@ -580,7 +580,7 @@ void DocHistory::OnScroll(wxScrollWinEvent& event) {
 }
 
 void DocHistory::OnVersionTreeSel(VersionTreeEvent& event) {
-	wxASSERT(event.GetItem() >= 0 && event.GetItem() < (int)m_items.size());
+	wxASSERT(event.GetItem() >= 0 && event.GetItem() < m_items.size());
 
 	// Do not return a reference. The events might modify
 	// m_items underneath it.
@@ -601,7 +601,7 @@ void DocHistory::OnVersionTreeSel(VersionTreeEvent& event) {
 }
 
 void DocHistory::OnVersionTreeTooltip(VersionTreeEvent& event) {
-	wxASSERT(event.GetItem() >= 0 && event.GetItem() < (int)m_items.size());
+	wxASSERT(event.GetItem() >= 0 && event.GetItem() < m_items.size());
 	const doc_id& hot_doc = m_items[event.GetItem()].doc;
 
 	cxLOCK_READ(m_catalyst)
@@ -626,7 +626,7 @@ void DocHistory::OnVersionTreeTooltip(VersionTreeEvent& event) {
 }
 
 void DocHistory::OnVersionTreeDrawitem(VersionTreeEvent& event) {
-	wxASSERT(event.GetItem() >= 0 && event.GetItem() < (int)m_items.size());
+	wxASSERT(event.GetItem() >= 0 && event.GetItem() < m_items.size());
 
 	const doc_id& hot_doc = m_items[event.GetItem()].doc;
 	int style = 0;
